# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Maia Learn API — implemented MVP
  version: 0.1.0
  description: Free article/Markdown/MP4 courses, resumable upload, private playback,
    curated home, identity and progress. HLS remains planned. Mutations require application/json
    except the authenticated octet-stream chunk route.
servers:
  - url: https://learn.maiaplatform.org/api/v1
  - url: http://localhost:3000/api/v1
paths:
  /auth/register:
    post:
      summary: Create a learner and queue verification email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                locale:
                  type: string
                  default: pt-BR
                  description: Saved account locale. The browser sends its selected interface locale. Account emails support en, pt-BR and es, with English fallback for other accepted locales.
      responses:
        '201':
          description: Account created
        '409':
          description: AUTH_EMAIL_EXISTS
        '422':
          description: Invalid fields
  /auth/login:
    post:
      summary: Rotate session and sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: User and session cookie
        '401':
          description: AUTH_INVALID_CREDENTIALS
  /auth/logout:
    post:
      summary: Destroy current session
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Logged out
  /auth/me:
    get:
      summary: Current user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User profile
        '401':
          description: Session missing or invalid
  /auth/verify-email:
    get:
      summary: Consume a verification token from a legacy email link
      deprecated: true
      description: Kept for previously queued or delivered email links. New emails link to the confirmation page at /auth/verify-email, which uses POST only after the user confirms.
      parameters:
        - name: token
          in: query
          required: true
          schema: {type: string, pattern: '^[a-f0-9]{64}$'}
      responses:
        '200': {description: Verified; code EMAIL_VERIFIED}
        '400': {description: AUTH_TOKEN_INVALID, AUTH_TOKEN_USED or AUTH_TOKEN_EXPIRED}
    post:
      summary: Explicitly confirm email ownership using a single-use verification token
      description: No session required. Token validation, consumption and user verification share one immediate transaction. Password-reset tokens cannot verify an email address. Existing verified_at timestamps are preserved.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: {type: string, pattern: '^[a-f0-9]{64}$'}
      responses:
        '200': {description: Verified; code EMAIL_VERIFIED}
        '400': {description: AUTH_TOKEN_INVALID, AUTH_TOKEN_USED or AUTH_TOKEN_EXPIRED}
        '429': {description: Authentication request limit reached}
  /auth/forgot-password:
    post:
      summary: Queue reset email in the saved account locale if the account exists
      description: Request locale is ignored. Known, unknown and missing addresses receive 204; the response does not reveal the account language or existence.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        '204':
          description: Accepted without disclosing account existence
  /auth/reset-password:
    post:
      summary: Consume reset token and invalidate existing sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - password
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
      responses:
        '200':
          description: Password updated
        '400':
          description: Invalid token
...
  /admin/categories:
    get:
      summary: List categories with version and references_count (administrator only)
      security:
        - cookieAuth: []
      responses:
        '200': {description: Category array; references_count includes all historical revisions}
        '403': {description: Administrator required}
    post:
      summary: Create a category (administrator only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '201': {description: Created category with id and version}
        '409': {description: CATEGORY_SLUG_EXISTS}
        '422': {description: Invalid category fields}
        '403': {description: Administrator required}
  /admin/categories/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: string, format: uuid}
    put:
      summary: Update global category text and slug immediately (administrator only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CategoryInput'
                - $ref: '#/components/schemas/CategoryVersion'
      responses:
        '200': {description: Updated category with incremented version}
        '404': {description: CATEGORY_NOT_FOUND}
        '409': {description: CATEGORY_CONFLICT or CATEGORY_SLUG_EXISTS}
        '422': {description: Invalid fields}
        '403': {description: Administrator required}
    delete:
      summary: Delete a category without references in any course revision (administrator only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryVersion'
      responses:
        '204': {description: Deleted}
        '404': {description: CATEGORY_NOT_FOUND}
        '409': {description: CATEGORY_CONFLICT or CATEGORY_IN_USE}
        '422': {description: Invalid expectedVersion}
        '403': {description: Administrator required}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-sid
      description: Production HTTPS cookie; local HTTP development uses sid.
...
  schemas:
...
    CategoryInput:
      type: object
      required: [slug, name]
      properties:
        slug: {type: string, minLength: 3, maxLength: 100, pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"}
        name: {type: string, minLength: 1, maxLength: 100}
        description: {type: string, maxLength: 1000, default: ''}
    CategoryVersion:
      type: object
      required: [expectedVersion]
      properties:
        expectedVersion: {type: integer, minimum: 1}
...